+++ src/models/analytics/analytics-tables.ts
@@ -112,6 +112,7 @@ type TableKeys =
   | 'gridPinGroup'
   | 'gridPin'
   | 'issue'
+  | 'issueAssignedUser'
   | 'measurement'
   | 'performanceMonth'
   | 'pin'
@@ -488,9 +489,9 @@ const analyticsTablesL10n: Record<TableKeys, AnalyticsTable> = {
       'hashId',
       'title',
       'level',
-      'typeKey',
-      'startAt',
-      'endAt',
+      'automation',
+      'assignedUserId',
+      'closedAt',
       'createdAt',
       'updatedAt',
       'deletedAt',
@@ -500,9 +501,9 @@ const analyticsTablesL10n: Record<TableKeys, AnalyticsTable> = {
         hashId: { singular: 'ID', plural: 'IDs' },
         title: { singular: 'Title', plural: 'Titles' },
         level: { singular: 'Priority', plural: 'Priorities' },
-        typeKey: { singular: 'Type', plural: 'Types' },
-        startAt: 'Started at',
-        endAt: 'Ended at',
+        automation: { singular: 'Automation type', plural: 'Automation types' },
+        assignedUserId: { singular: 'Assigned user ID', plural: 'Assigned user IDs' },
+        closedAt: 'Closed at',
         createdAt: 'Created at',
         updatedAt: 'Updated at',
         deletedAt: 'Deleted at',
@@ -514,6 +515,9 @@ const analyticsTablesL10n: Record<TableKeys, AnalyticsTable> = {
         typeKey: { singular: 'Type', plural: 'Types' },
         startAt: 'Begon op',
         endAt: 'Eindigde op',
+        automation: { singular: 'Automatisering type', plural: 'Automatisering types' },
+        assignedUserId: { singular: 'Toegewezen aan ID', plural: 'Toegewezen aan IDs' },
+        closedAt: 'Gesloten op',
         createdAt: 'Aangemaakt op',
         updatedAt: 'Bijgewerkt op',
         deletedAt: 'Verwijderd op',
@@ -523,6 +527,24 @@ const analyticsTablesL10n: Record<TableKeys, AnalyticsTable> = {
     tableText: {
       en: { singular: 'Issue', plural: 'Issues' },
       nl: { singular: 'Issue', plural: 'Issues' },
+    },
+  },
+  issueAssignedUser: {
+    fields: ['hashId', 'name'],
+    fieldsWithTranslations: {
+      en: {
+        hashId: { singular: 'ID', plural: 'IDs' },
+        name: { singular: 'Name', plural: 'Names' },
+      },
+      nl: {
+        hashId: { singular: 'ID', plural: 'IDs' },
+        name: { singular: 'Naam', plural: 'Namen' },
+      },
+    },
+    tableKey: 'issueAssignedUser',
+    tableText: {
+      en: { singular: 'Issue assigned user', plural: 'Issue assigned users' },
+      nl: { singular: 'Issue toegewezen aan', plural: 'Issue toegewezen aan' },
     },
   },
   measurement: {