+++ src/routes/issue/find.ts
@@ -11,6 +11,7 @@ interface Query extends TableQuery {
   pinGroupHashId?: string | null;
   edgeHashId?: string | null;
   gridHashId?: string | null;
+  labelHashId?: string | null;
   mapLayers?: string[] | null;
 }
 
@@ -22,6 +23,8 @@ interface EffectiveQuery extends EffectiveTableQuery {
   pinGroupHashId: string | null;
   edgeHashId: string | null;
   gridHashId: string | null;
+  labelHashId: string | null;
+  mapLayers: string[] | null;
 }
 
 interface EffectiveRequest {
@@ -34,7 +37,7 @@ interface ResponseRow {
   assignedUserName: string | null;
   pinGroup: PinGroup;
   subscribed: boolean;
-  links: { // TODO withthegrid/platform#1586 Rob: deprecated
+  links?: { // TODO withthegrid/platform#1586 Rob: deprecated
     pinGroup: PinGroup;
     pin: Pin | null;
   }[];
@@ -53,6 +56,7 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
       pinGroupHashId: Joi.string().allow(null).default(null),
       edgeHashId: Joi.string().allow(null).default(null),
       gridHashId: Joi.string().allow(null).default(null),
+      labelHashId: Joi.string().allow(null).default(null),
       mapLayers: Joi.array().items(Joi.string()).allow(null).default(null),
     }),
   right: { environment: 'READ' },
@@ -70,7 +74,7 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
       links: Joi.array().min(1).items(Joi.object().keys({
         pinGroup: pinGroupSchema.required(),
         pin: pinSchema.allow(null).required(),
-      })).required(),
+      })),
     })).required(),
   }).required(),
   description: 'Search through issues',